RESKC-980:Do not display certification toggle functionality when Prop Person COI Status param is enabled
The COI Disclosure Status flag feature is designed to control if the proposal person certification questionnaire should appear for proposal personnel.
When the contribution was initially merged, it 'broke' the manual 'include certification questions' button that appears for Key Persons. RESKC-688 restored that manual option.
This case is to hide the manual 'include certification questions' button when the COI Disclosure Status feature is enabled. (PROP_PERSON_COI_STATUS_FLAG = Y)

The contributed feature provided an override mechanism (custom attribute) that lets the user apply the PPC questionnaire if the sponsor is not in the COI Hierarchy, so this button is not needed.