RESKC-704:Subaward Obligated Fund modifications do not update associated Award record
As an Award and Subaward user, I reference the Award module > Subaward where this award is a funding source panel to confirm the funding amount obligated to those records.
Currently, the Award record will successfully update with the initial subaward fund allocation. But successive modifications that alter the funding level of the subaward are not reflected in the award.