RESMER-457:Subaward: Funding Source panel does not reflect information as it is updated in the Award.
I noticed that the Funding Source panel > Current Funding Sources subpanel in the Subaward tab does 'refresh' the information from the most recent Award Version (instead the info is static). Also, when a user clicks the Open award link (in the Funding Source panel of Subaward tab), the system opens up the Award Version document that was available at the time this award was added to the Funding Source panel of the Subaward Document and NOT the latest finalized Award Version that is available in the system now.