RESMER-452: Subaward: System generates wrong Subaward version numbers when not the latest Subaward Document is selected to be edited.
The system is not generating the Subaward Versions correctly. If, for example, my Subaward has 4 Versions and I open and edit the Subaward Document that corresponds to Version 1 (searching for it in the doc search). The system will assign the new Subaward Document version 2 as opposed to version 5. It's assigning the next version number from the Document opened and not the next version number in the whole sequence. This is also causing the system to create duplicate/multiple Subawards with the same Subaward ID.