RESMER-459:Subaward: Last Update field in the Document Header does not display the correct Username once the Subaward Document is in Final status
In the Subaward Document Header, the Last Update field displays the wrong username (of the user that made the update/change) when the document status is set to 'FINAL'. Instead, the system displays user 'kr' as the one who made the Last Update. When the Document is in 'SAVED' status the correct user name displays.