RESMER-462:PD - Inactive COI Notification based on answers still generating
Confirmed the PD notification 10407 "Notification based on answers to the COI screening questions" was not Active in MITQA.
Created PD 30523 added personnel. Backdoor logged in as co-Investigator user mgb who answered Yes to all three COI screening questions. We got the on screen prompt modal instructing user to disclose in Coeus lite. And the attached email notification was generated to mgb stating user must complete a full disclosure before proposal can be routed.

Since this notification is not active it should not generate.