RESKC-1091:Proposals with Leap Year Feb 1 start date mishandle budget autocalculation period start day in non-leap years
If a proposal & thus budget has a February 1 start date in a Leap Year, the autocalculate tool mishandles the non-leap year start date for some personnel.
Specifically, if the person has a row that is not assigned start on 2/1, in non-leap years, that row starts on the 2nd of the month.