RESKC-1095:PD Budget: new validation to enforce Proposal and Budget Start/End dates breaks several business processes
The current fix does not allow the flexibility needed for different types of proposal submissions to comply with sponsor requested preparation (e.g. supplement proposals where the entire project period is to be entered in the proposal but the budget is to include on the start and end date covering the supplemental funds spending), or for syncing child proposal budget date changes to a hierarchy.

The validation needs to be changed from requiring budget start and end dates to match the proposal start and end dates to instead enforcing that the budget start date falls after the proposal start date and the budget end date falls before the proposal period end date.
The budget start and end dates must both fall within the project period, rather than match exactly.