RESOPS-577:Invalid field constraints prevent Award save when amounts are negative
When a T&M transaction includes negative amounts and the indirect flag is enabled, incorrect character constraints would prevent an award from being saved.
Additionally fixing saving amounts in the award itself, in particular when the award is new. This ensures the distributable amounts and change values are correct even when entered on the award home screen.