RESKC-572:fix period 2 inflation calculation
This is not fixed in demo 1 either. And all calculations are impacted in a negative way.
P1 - All salaries are correctly calculating for period 1 to reflect anniversary or regular inflation start date.
P2: Anniversary salary and regular salary inflation are not being inflated in period 2.
P3: 9 and 10 month appointments begin inflation, amounts that should have appeared in p2.
But the 12 month appointments vary; anniversary rate 'caught up' and correct for the period, but the regular inflation salary is at what should have been p2.
P4: inconsistent rate amount applied: no amounts match the expected salary
P's5, 6 & 7: SOME salaries match the projected amount; But the 9 and 10 month Anniversary rates salaries display what was expected for the prior period.
Ps 8, 9, and 10: most of the salary calculations are incorrect, with several examples REDUCING amounts from the prior period.