RESKC-883: fixing number proposal hierarchy issues.
1 - When hierarchy parent status is Approval Pending, and children are displayed as such, when child is opened with the View link in the PD search results Actions link, the following are available:
Edit button
Save button
From top menu Hierarchy > Sync Hierarchy and Unlink Hierarchy
Child proposals that are routing should be locked down and NOT able to be edited, synced to parent or unlinked.
2 - When hierarchy parent status is Approved and Submitted, and children are displayed as such, when child is opened with the View link in the PD search results Actions link, the following are available:
Edit button
Save button
From top menu Hierarchy > Sync Hierarchy and Unlink Hierarchy
Child proposals that are Approved and Submitted should NOT be able to be edited Synced, linked or unlinked.
3)
log on with quickstart (or some other user)
create a proposal with budget
create hierarchy
do not close child with close button or manually delete any pessimistic locks but use proposal search or doc search to find the parent proposal
open parent proposal, fill in required fields, submit
with the same user, in the same browser session, open the child proposal
Notice that the child proposal is fully editable but should not be. This is because the child has a pessimistic lock and rice thinks that if a user has a lock then the user can edit.