RESMER-531:When no stop is created during reject, the PD continues workflow immedietly after reject.
When this happen the default workflow falls to the KRMS Agendas and Peopleflow immedietly meaning that changes to the proposal during revisions requested will not affect the KRMS rules for routing and approvals.

This change adds an approval stop for all current document level Aggregators. It also allows any user with the permission to submit the proposal to deactivate those approval requests so the proposal will continue into workflow.