RESKC-1117: Adding new KRMS functions for custom data.
Need a new function for KRMs in all Namespaces to search for a custom data entry. We're imagining this could work in a similar manner to the Questionnaire Category.
For Parameters, user may need to specify the specific Custom Attribute ID and the Document Type Code so that the specific field in question can be searched.
*Needs to work for Comparison of Null and !=Null, *
Would be nice to work where it can reference specific Value also