RESKC-1132: Fixed proposal development document when recalled.
Proposal Status is "Approval Pending" instead of "Revisions Requested" on proposals recalled during routing
Proposal Routed for approval and was Recalled before the Investigator approved.
When in the same session, the submit status changed to save and the document became editable, user navigated to the key personnel section, searched for and added a multi-pi. selected the person certifications and noticed that answers to questions were not editable.
Left proposal and searched for proposal in proposal search.
Can edit some areas of proposal persons (such as commons name) but cannot certify the newly added multipi. Document status shows as approval pending and submit screen shows as saved (not routing)
Route log shows that the initiator recalled the proposal
yet the proposal is not entirely editable and the document status is incorrectly approval pending
To reproduce:
Searched for Proposal enroute that was awaiting PI approval
Clicked Edit from Search Results
Went to Submit on left Nav
Recalled
added Recall reason
navigated to key persons
added key person to Multi PI role
Closed.
Searched for PD again
Edit proposal
proposal shows as approval in progress and is not completely editable