RESKC-1203: Recalculating budget periods if posted budget exists.
On an award we are not able to create a new budget revision because it gives us an error "Period 2 end date cannot exceed project end date" But we can not change this in any way. The project was shortened after the previous budget was posted but now it will not allow us to rebudget using the new date. Please advise as how we should accomplish this and why this is a problem.
steps to reproduce
Create an award,
create an award budget,
approve, post.
Via T&M document,
change award dates
now try to create a new budget, get error message.