RESKC-1165, RESKC-1048: Adding Coi disclosure status to IRB and IACUC protocol persons.
As a researcher or admin responsible for a protocol, I can see the COI disclosure status of project person to determine if the protocol can be approved as is or or if project personnel need to disclose to proceed.
General Acceptance Criteria:
Configurable in monolith so it is only displayed in a module if desired by the school.
For each project person, display the latest status of the annual disclosure that includes that project.
if the project is not on annual yet, display "Not yet disclosed"
if a project is not required to disclose display "Disclosure Not Required"
Acceptance Criteria:
Admin can configure IRB to link with new COI module
COI disclosure status for protocol project persons displays in IRB protocol based new module
Design: display "COI Disclosure Status: [status]" in Person header