RESKC-1211: IRB and IACUC FYI process enhancement
Problem: IRB was not always easily able to keep track of all notify IRB (FYI) actions, because admin only needs to acknowledge. This does not support a process where IRB admins often need to go back and forth with PI/Dept Admin to get needed documentation.
Contributed Solution:
-Model a special routing for the FYI/ Notify IRB process to be similar to the amendment protocol process, except does not allow modification of the protocol itself other than notes and attachments.
-IRB can now obtain more information and even approve/disapprove action.
-Upon approve the attachments role into the notes and attachments history for tracking.
-Parameterized.
-IACUC is also addressed.
This enhancement was discussed and approved by the KCC IRB subcommittee in 05/2014. Foundation Jira is KRAFDBCK-10691. Also, was shared with the IACUC Subcommittee.