RESKC-673: fix footer getting stuck in middle of page
Losing access to some of the left-hand navigation buttons, including Delivery Info, Budget, Access, Summary/Submit, etc. when page footer floats and does not allow user to scroll past.
I believe it IS happening consistently when both left nav sections that have sub sections are open; so when both Basics and Key Personnel are open and showing sub menu options.
Only workaround we have identified is to keep clicking nav buttons toward the top of the list which refreshes the screen until the issue resolves.