RESKC-1488: update page when closing budget versions dialog
Create a Summary budget
Enter a direct cost amount of any value
Toolbar > Budget Versions > Action > Complete Budget. OK to complete . Okay to close the modal.
Result: 'complete budget' button is still present: Budget is NOT locked down until user takes an action that causes a save.
1. clicks save
2. navigates to another window
3. opens Budget Settings and closes
3. clicks Complete Budget
ISSUE: When the Periods & Totals screen is in focus and this Toolbar "complete budget' option is used, the P&T screen remained Editable. I was able to revise the data on that screen, and then at Save, it was saved the data and finalized the Complete Budget action WITH That change.