RESKC-1838: Fix hierarchy line item group conflict. (#2095)
When viewing personnel item Details & Rates, when the Default or Custom Group includes entries from multiple children, the Details & Rates only reflects the amounts from the child budget that created the parent. The info from the other personnel items are not reflected. (This seems to be happening in all hierarchy proposals - not just the ones with LA and non-LA units)
(Only if each child budget has separate Custom Group(s) they used for entering each of their shared person cost element types, then it shows correctly in the details & rates. Or, if there is a Default or Custom group (g/l or cost element) that contains only salaries from one child, and there are no personnel from the other children in that cost element/category, then it displays correctly.)